Fix: 기존 XML -> java Config 변경 과정에서 나뉘어 져 있던 설정 기능 통합.
기존 XML -> java Config 변경 과정에서 나뉘어 져 있던 EgovConfigApp.java 와 EgovConfigAppCommon.java를 EgovConfigApp.java로 통합.
공통 설정은 앞으로 EgovConfigApp.java를 통해 설정하면 됨.
